refactor(backend): Reorganize & clean up execution models & DB functions
